requirements.txt
Flask==3.0.3
supabase==2.4.5
flask-bcrypt==1.0.1
python-dotenv==1.0.1
flask-cors==4.0.1
